refactor(autoconfig): consolidate LLMRefitPolicy onto shared ConfigEdit vocabulary
Extracts the ConfigEdit lever vocabulary into a new cycle-free module
core/config_edits.py (the optimizer imports the policy, so the shared vocabulary
can't live in either). Both the optimizer's proposers and the controller's
LLMRefitPolicy now import edits from there.

LLMRefitPolicy's LLM repair now emits the same closed edit vocabulary as the
optimizer ({"edits": [{"op": ...}]}, parsed by parse_llm_edits and folded onto the
current config by fold_edits) instead of raw config diffs, so the controller's
single-trajectory repair and the optimizer's search speak one lever language. The
parse+fold step is factored into _config_from_edits for testing without the openai
dep. apply_config_diff is retained as a legacy helper.

LLM repair is env-gated (GOLDENMATCH_AUTOCONFIG_LLM=1, default off) so default
auto-config output is unchanged. 143 optimizer/autoconfig/zero-label tests green.
